BaseStation/apiwebapp/requirements.txt
amqp==2.1.4
anyjson==0.3.3
appnope==0.1.0
asyncio==3.4.3
asynqp==0.5.1
billiard==3.5.0.2
blessings==1.6
bpython==0.16
celery==4.0.2
certifi==2017.4.17
chardet==3.0.4
curtsies==0.2.11
cycler==0.10.0
decorator==4.0.11
Django==1.11.2
djangorestframework==3.6.4
enum-compat==0.0.2
eventlet==0.21.0
greenlet==0.4.12
idna==2.5
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
kombu==4.0.2
matplotlib==2.0.0
mock==2.0.0
mysqlclient==1.3.10
nameko==2.6.0
nose==1.3.7
numpy==1.13.1
olefile==0.44
path.py==10.3.1
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
pika==0.10.0
Pillow==4.0.0
prompt-toolkit==1.0.14
ptyprocess==0.5.2
pyclustering==0.6.6
Pygments==2.2.0
pyparsing==2.1.10
pyserial==3.3
pyserial-asyncio==0.4
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
requests==2.18.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
urllib3==1.21.1
vine==1.1.3
wcwidth==0.1.7
Werkzeug==0.12.2
wrapt==1.10.10
XBee==2.2.5
yagmail==0.9.178
